fix(spaces): do not truncate the global DoF count to int
DiscontinuousMapper::size() accumulated size_t summands into the int
literal 0, truncating/overflowing the global DoF count for spaces with
more than 2^31 DoFs.

See docs/reviews/2026-07-adversarial-cpp-review.md, Section B (PR #305).

+++ dune/gdt/spaces/mapper/discontinuous.hh
@@ -92,7 +92,7 @@ public:
 
   size_t size() const override final
   {
-    return std::accumulate(size_.begin(), size_.end(), 0);
+    return std::accumulate(size_.begin(), size_.end(), size_t(0));
   }
 
   size_t max_local_size() const override final
